% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdockerversion.R
\name{getRNAdockerversion}
\alias{getRNAdockerversion}
\title{getdockerversion}
\usage{
getRNAdockerversion()
}
\value{
RNAdocker版本号
}
\description{
输出RNA和WES流程docker版本号
}
\details{
cp
}
\examples{
no example

}
